Update _variables.tf
Warning: Quoted type constraints are deprecated

  on .terraform/modules/client-vpn/_variables.tf line 20, in variable "tags":
  20:   type        = "map"

Terraform 0.11 and earlier required type constraints to be given in quotes,
but that form is now deprecated and will be removed in a future version of
Terraform. To silence this warning, remove the quotes around "map" and write
map(string) instead to explicitly indicate that the map elements are strings.